+++ signature/src/armv7/dilithium_ntt_consts.h
@@ -19,7 +19,7 @@
 
 /*
  * This code is derived in parts from the code distribution provided with
- * git@github.com:dsprenkels/dilithium-cortexm.git
+ * https://github.com/dilithium-coretexm/dilithium-cortexm.git
  *
  * author: Markus Krausz
  * date: 18.03.18