+++ signature/src/armv8/dilithium_signature_armv8.c
@@ -24,6 +24,7 @@
 #define LC_POLY_UNIFOR_BUF_SIZE_MULTIPLIER 2
 
 #include "dilithium_poly.h"
+#include "dilithium_poly_common.h"
 #include "dilithium_poly_armv8.h"
 #include "dilithium_polyvec.h"
 #include "dilithium_polyvec_armv8.h"