+++ signature/src/dilithium_signature_c.c
@@ -32,6 +32,7 @@
 #define LC_POLY_UNIFOR_BUF_SIZE_MULTIPLIER 1
 
 #include "dilithium_poly.h"
+#include "dilithium_poly_common.h"
 #include "dilithium_poly_c.h"
 #include "dilithium_polyvec.h"
 #include "dilithium_polyvec_c.h"